Refac/rename hiveboard to hivemind (#21)
* Rename cpp files

* Rename classes and contents of files

* format

* rename ros package

* rename include folder

* rename hivemind in cmakelists

* change launchfiles contents

* gitignore
