+++ src/hive_mind_bridge/cmake/propolis.cmake
