+++ src/hive_mind_bridge/launch/hive_mind_bridge.launch
@@ -1,0 +1,7 @@
+<?xml version="1.0" ?>
+<launch>
+    <node pkg="hive_mind_bridge" type="hive_mind_bridge" name="hive_mind_bridge" output="screen">
+        <param name="TCP_SERVER_PORT" type="int" value="8080"/>
+        <param name="ROBOT_NAME" type="str" value="pioneer_0"/>
+    </node>
+</launch>
