+++ src/hive_mind_bridge/launch/socket_to_move_by_integration_test.launch
@@ -1,0 +1,10 @@
+<?xml version="1.0" ?>
+<launch>
+    <node pkg="hive_mind_bridge" type="hive_mind_bridge" name="hive_mind_bridge" output="screen">
+        <param name="TCP_SERVER_PORT" type="int" value="8080"/>
+        <param name="ROBOT_NAME" type="str" value="pioneer_0"/>
+    </node>
+
+    <node pkg="hive_mind_bridge" type="SocketToMoveByIntegrationTest" name="hive_mind_bridge_integration_test" output="screen"/>
+
+</launch>
