+++ src/hive_mind_bridge/test/mocks/HiveMindHostSerializerInterfaceMock.h
@@ -1,0 +1,14 @@
+#ifndef HIVEMIND_BRIDGE_HIVEMINDHOSTSERIALIZERINTERFACEMOCK_H
+#define HIVEMIND_BRIDGE_HIVEMINDHOSTSERIALIZERINTERFACEMOCK_H
+
+#include <gmock/gmock.h>
+#include <hivemind-host/IHiveMindHostSerializer.h>
+
+class HiveMindHostSerializerInterfaceMock : public IHiveMindHostSerializer {
+  public:
+    ~HiveMindHostSerializerInterfaceMock() = default;
+
+    MOCK_METHOD((bool), serializeToStream, (const MessageDTO& message), (override));
+};
+
+#endif // HIVEMIND_BRIDGE_HIVEMINDHOSTSERIALIZERINTERFACEMOCK_H
