--- src/hiveboard_bridge/include/hiveboard_bridge/MessageHandler.h
@@ -1,22 +1,0 @@
-#ifndef CATKIN_ROS_MESSAGEHANDLER_H
-#define CATKIN_ROS_MESSAGEHANDLER_H
-
-#include "IMessageHandler.h"
-#include "hiveboard_bridge/MessageUtils.h"
-
-class MessageHandler : public IMessageHandler {
-  public:
-    MessageHandler();
-    ~MessageHandler();
-
-    MessageDTO handleMessage(MessageDTO message) override;
-
-    bool registerCallback(std::string name, CallbackFunction callback) override;
-
-    std::optional<CallbackFunction> getCallback(const std::string& name) override;
-
-  private:
-    CallbackMap m_callbacks;
-};
-
-#endif // CATKIN_ROS_MESSAGEHANDLER_H
