--- src/hiveboard_bridge/launch/hiveboard_bridge.launch
@@ -1,7 +1,0 @@
-<?xml version="1.0" ?>
-<launch>
-    <node pkg="hiveboard_bridge" type="hiveboard_bridge" name="hiveboard_bridge" output="screen">
-        <param name="TCP_SERVER_PORT" type="int" value="8080"/>
-        <param name="ROBOT_NAME" type="str" value="pioneer_0"/>
-    </node>
-</launch>
