--- src/hiveboard_bridge/launch/socket_to_move_by_integration_test.launch
@@ -1,10 +1,0 @@
-<?xml version="1.0" ?>
-<launch>
-    <node pkg="hiveboard_bridge" type="hiveboard_bridge" name="hiveboard_bridge" output="screen">
-        <param name="TCP_SERVER_PORT" type="int" value="8080"/>
-        <param name="ROBOT_NAME" type="str" value="pioneer_0"/>
-    </node>
-
-    <node pkg="hiveboard_bridge" type="SocketToMoveByIntegrationTest" name="hiveboard_bridge_integration_test" output="screen"/>
-
-</launch>
