--- src/hiveboard_bridge/src/HiveBoardBridge.cpp
@@ -1,16 +1,0 @@
-#include "hiveboard_bridge/HiveBoardBridge.h"
-
-HiveBoardBridge::HiveBoardBridge(int tcpPort) :
-    m_tcpServer(tcpPort), m_deserializer(m_tcpServer), m_serializer(m_tcpServer) {
-    m_bridge = std::make_unique<HiveBoardBridgeImpl>(m_tcpServer, m_serializer, m_deserializer);
-}
-
-void HiveBoardBridge::spin() { m_bridge->spin(); }
-
-void HiveBoardBridge::onConnect(std::function<void()> hook) { m_bridge->onConnect(hook); }
-
-void HiveBoardBridge::onDisconnect(std::function<void()> hook) { m_bridge->onDisconnect(hook); }
-
-bool HiveBoardBridge::registerCustomAction(std::string name, CallbackFunction callback) {
-    return m_bridge->registerCustomAction(name, callback);
-}
